requirements.txt
Django==3.2.12
djangorestframework==3.13.1
django-crontab==0.7.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-filter==21.1
apachelogs==0.6.0
